refactor: domain.utils.Logger -> domain.helper.Logger.
