+++ app/src/main/java/no/usn/mob3000/ui/components/game/board/ChessBoard.kt
@@ -32,7 +32,7 @@ import com.github.bhlangonijr.chesslib.Piece
 import com.github.bhlangonijr.chesslib.Rank
 import com.github.bhlangonijr.chesslib.Side
 import com.github.bhlangonijr.chesslib.Square
-import no.usn.mob3000.domain.utils.Logger
+import no.usn.mob3000.domain.helper.Logger
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
