+++ app/src/main/java/no/usn/mob3000/ui/screens/chess/PlayScreen.kt
@@ -14,29 +14,22 @@ import androidx.compose.material3.Card
 import androidx.compose.material3.CardDefaults
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableIntStateOf
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.StateFlow
 import no.usn.mob3000.R
 import no.usn.mob3000.domain.helper.game.convertPgnToFen
-import no.usn.mob3000.domain.model.game.board.ChessBoardEvent
 import no.usn.mob3000.domain.model.game.board.ChessBoardGameState
 import no.usn.mob3000.domain.model.game.board.ChessBoardState
 import no.usn.mob3000.domain.model.game.board.PracticeMode
 import no.usn.mob3000.domain.model.game.opening.Opening
-import no.usn.mob3000.domain.utils.Logger
-import no.usn.mob3000.domain.viewmodel.game.ChessBoardViewModel
+import no.usn.mob3000.domain.helper.Logger
 import no.usn.mob3000.ui.components.base.Viewport
 import no.usn.mob3000.ui.components.game.board.ChessBoard
 import no.usn.mob3000.ui.theme.DefaultButton
